requirements.txt
requests==2.28.1
selenium==4.7.2
pytest==7.2.0
allure-pytest==2.12.0
lxml==4.9.1
flake8==6.0.0
